requirements.txt
Flask==2.0.2
tensorflow==2.17.0
keras==3.4.1
h5py==3.11.0
numpy==1.26.4
Pillow==9.4.0
requests==2.26.0
Werkzeug==2.0.3
imageio==2.31.1